Define page_map_entry_t

+++ virtual-memory/memory.h
@@ -52,16 +52,15 @@ typedef struct {
   uint32_t swap_loc;
   uint32_t vaddr;     /* virtual address */
   uint32_t ppage_num; /* physical page number */
-  // TODO: use bit fields to avoid use of masks
-  // uint32_t bits;
+
   /* page directory/table entry bits */
-  uint32_t p: 1;
-  uint32_t rw: 1;
-  uint32_t us: 1;
-  uint32_t pwt: 1;
-  uint32_t pcd: 1;
-  uint32_t a: 1;
-  uint32_t d: 1;
+  uint8_t p: 1;   /* present */
+  uint8_t rw: 1;  /* read/write */
+  uint8_t us: 1;  /* user/supervisor */
+  uint8_t pwt: 1; /* page write-through */
+  uint8_t pcd: 1; /* page cache disable */
+  uint8_t a: 1;   /* accessed */
+  uint8_t d: 1;   /* dirty */
 } page_map_entry_t;
